Add redis database to the scheduler
This commit, if applied, adds a redis db to the scheduler.

In order to be able to scale the number of replicas of the scheduler, we need a
way to synchronize the processes. This is can be achieved by using a database
like redis.

Another thing to consider that will require a DB is the future feature:

  taints and tolerations

Taints applied to clouds will be dynamic resources of the scheduler.
The different modules (prometheus, calendar, etc) will possibily taint a cloud,
to reduce its priority, or to set it as unschedulable. An Operator should also
be able to taint a cloud with a single API call.

Again, how we are going to synchronize the taints between the replicas, plus the
ability to recover the state of the taints when the pods are restarted, will be
possible with a DB.