Fix update of the configuration from Git repository
This commit, if applied, fixes the /api/v1/repo PUT endpoint.

The git WorkTree method Pull() was not working because the git worktree didn't have any
history because of this setting in the CloneOptions:

  Depth: 1

Removing that makes the worktree able to Pull from remote.

Also specify SingleBranch, as we don't work with multiple branches.